libretro: Fix whitespace issues, integrate libretro_wrapper into libretro_ui
