+++ src/libretro_ui.h
@@ -1,15 +1,34 @@
-#ifndef _EASYRPG_LIBRETROUI_H_
-#define _EASYRPG_LIBRETROUI_H_
+/*
+ * This file is part of EasyRPG Player.
+ *
+ * EasyRPG Player is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * EasyRPG Player is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with EasyRPG Player. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef EP_LIBRETRO_UI_H
+#define EP_LIBRETRO_UI_H
+
+#ifdef USE_LIBRETRO
 
 // Headers
+#include "audio_libretro.h"
 #include "baseui.h"
 #include "color.h"
 #include "rect.h"
 #include "system.h"
 
-#ifdef USE_LIBRETRO
 #include "libretro.h"
-#endif
+
 /**
  * LibretroUi class.
  */
@@ -26,15 +45,9 @@ public:
 	LibretroUi(int width, int height);
 
 	/**
-	 * Destructor.
-	 */
-	~LibretroUi();
-
-	/**
 	 * Inherited from BaseUi.
 	 */
 	/** @{ */
-
 	void BeginDisplayModeChange() override;
 	void EndDisplayModeChange() override;
 	void Resize(long width, long height) override;
@@ -50,32 +63,30 @@ public:
 
 	uint32_t GetTicks() const override;
 	void Sleep(uint32_t time_milli) override;
+
 #ifdef SUPPORT_AUDIO
-	AudioInterface& GetAudio();
+	AudioInterface& GetAudio() override;
 	std::unique_ptr<AudioInterface> audio_;
 #endif
-
 	/** @} */
-#ifdef USE_LIBRETRO
+
 #if defined(HAVE_OPENGL) || defined(HAVE_OPENGLES)
 public:
 	static void ResetRetroGLContext(void);
 	static void DestroyRetroGLContext(void);
 	static bool LockRetroGLFramebuffer(void *data);
-private:	
+private:
 	static bool retro_gl_framebuffer_ready;
-	
 #endif
+
 public:
 	static void SetRetroVideoCallback(retro_video_refresh_t cb);
-	static void SetRetroInputStateCallback(retro_input_state_t cb); 
+	static void SetRetroInputStateCallback(retro_input_state_t cb);
 	static retro_usec_t time_in_microseconds;
 private:
 	static retro_video_refresh_t UpdateWindow;
 	static retro_input_state_t CheckInputState;
-	
-#endif
-	
+
 };
 
 #endif
